+++ molten_jwt/exceptions.py
@@ -1,4 +1,3 @@
-from typing import Dict, Any
 from molten.errors import MoltenError
 
 
@@ -16,8 +15,3 @@ class AuthorizationError(MoltenError):
 
 class TokenValidationError(MoltenError):
     """To be raised when a JWTClaims validation check fails"""
-    def __init__(self, reasons: Dict[str, Any]) -> None:
-        self.reasons = reasons
-
-    def __str__(self) -> str:
-        return str(self.reasons)